requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0
django-admin-generator==2.6.0
django-environ==0.11.2
django-simple-math-captcha==2.0.1
django-utils2==3.0.2
gunicorn==22.0.0
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
python-utils==3.8.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
